requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.3
celery[redis]==5.6.2
redis==6.2.0
pydantic==2.11.4
pydantic-settings==2.9.1
python-multipart==0.0.20
transformers>=4.51.0,<4.58.0
diffusers>=0.33.0
accelerate>=1.6.0
safetensors>=0.5.3
huggingface-hub>=0.28.0
soundfile>=0.13.1
librosa>=0.10.2
scipy>=1.15.2
einops>=0.8.1
vector-quantize-pytorch>=1.24.6
